Manejo de errores en Java y Angular

+++ eclipse-workspace/spring-boot-backend-apirest/src/main/java/com/crisperez/springboot/backend/apirest/controllers/ClienteRestController.java
@@ -1,10 +1,13 @@
 package com.crisperez.springboot.backend.apirest.controllers;
 
-import java.util.Date;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -32,31 +35,94 @@ public class ClienteRestController {
 	}
 	
 	@GetMapping("/clientes/{id}")
-	public Cliente show(@PathVariable Long id) { // Viene en la URL, por eso PathVariable
-		return clienteService.findById(id);
+	public ResponseEntity<?> show(@PathVariable Long id) { // Viene en la URL, por eso PathVariable
+		Cliente cliente = null;
+		Map<String, Object> response = new HashMap<>();
+
+		try {
+			cliente = clienteService.findById(id);
+		}catch(DataAccessException e) {
+			response.put("mensaje", "Error al realizar la consulta en la base de datos.");
+			response.put("error", e.getMessage().concat(": ").concat(e.getMostSpecificCause().getMessage()));
+			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+		
+		if(cliente==null) {
+			response.put("mensaje", "El cliente ID: ".concat(id.toString().concat(" no existe.")));
+			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.NOT_FOUND);
+		}
+		
+		return new ResponseEntity<Cliente>(cliente, HttpStatus.OK);
 	}
 	
 	@PostMapping("/clientes")
-	@ResponseStatus(HttpStatus.CREATED)
-	public Cliente create(@RequestBody Cliente cliente) { // Viene en el Body del request, como JSON
-		return clienteService.save(cliente);
+	public ResponseEntity<?> create(@RequestBody Cliente cliente) { // Viene en el Body del request, como JSON
+		Cliente nuevoCliente = null;
+		Map<String, Object> response = new HashMap<>();
+		
+		try {
+			nuevoCliente = clienteService.save(cliente);
+		}catch(DataAccessException e) {
+			response.put("mensaje", "Error al realizar insert en la base de datos.");
+			response.put("error", e.getMessage().concat(": ").concat(e.getMostSpecificCause().getMessage()));
+			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+		
+		response.put("mensaje", "El cliente ha sido creado con éxito.");
+		response.put("cliente", nuevoCliente);
+		return new ResponseEntity<Map<String, Object>>(response, HttpStatus.CREATED);
 	}
 	
 	@PutMapping("/clientes/{id}")
-	@ResponseStatus(HttpStatus.CREATED)
-	public Cliente update(@RequestBody Cliente cliente, @PathVariable Long id) {
-		Cliente clienteActual = clienteService.findById(id);
+	public ResponseEntity<?> update(@RequestBody Cliente cliente, @PathVariable Long id) {
+		Cliente clienteActual = null;
+		Map<String, Object> response = new HashMap<>();
+		
+		try {
+			clienteActual = clienteService.findById(id);
+		}catch(DataAccessException e) {
+			response.put("mensaje", "Error al realizar update en la base de datos.");
+			response.put("error", e.getMessage().concat(": ").concat(e.getMostSpecificCause().getMessage()));
+			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+		
+		if(clienteActual==null) {
+			response.put("mensaje", "El cliente ID: ".concat(id.toString().concat(" no existe.")));
+			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.NOT_FOUND);
+		}
 		
 		clienteActual.setApellido(cliente.getApellido());
 		clienteActual.setNombre(cliente.getNombre());
 		clienteActual.setEmail(cliente.getEmail());
+		clienteActual.setCreateAt(cliente.getCreateAt());
 		
-		return clienteService.save(clienteActual);
+		clienteActual = clienteService.save(clienteActual);
+		response.put("mensaje", "El cliente ha sido actualizado con éxito.");
+		response.put("cliente", clienteActual);
+		
+		
+		return new ResponseEntity<Map<String, Object>>(response, HttpStatus.ACCEPTED);
 	}
 	
 	@DeleteMapping("/clientes/{id}")
-	@ResponseStatus(HttpStatus.NO_CONTENT)
-	public void delete(@PathVariable Long id) {
-		clienteService.delete(id);
+	public ResponseEntity<?> delete(@PathVariable Long id) {
+		Map<String, Object> response = new HashMap<>();
+		Cliente clienteEliminable = clienteService.findById(id);;
+		
+		if(clienteEliminable==null) {
+			response.put("mensaje", "El cliente ID: ".concat(id.toString().concat(" no existe.")));
+			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.NOT_FOUND);
+		}
+		
+		try {
+			clienteService.delete(id);
+		}catch(DataAccessException e) {
+			response.put("mensaje", "Error al eliminar el cliente en la base de datos.");
+			response.put("error", e.getMessage().concat(": ").concat(e.getMostSpecificCause().getMessage()));
+			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+		
+		response.put("mensaje", "Cliente eliminado correctamente");
+		return new ResponseEntity<Map<String, Object>>(response, HttpStatus.OK);
 	}
 }
